// interfaces/Uf.ts
export interface Uf {
  name:
    | "Acre"
    | "Alagoas"
    | "Amapá"
    | "Amazonas"
    | "Bahia"
    | "Ceará"
    | "Distrito Federal"
    | "Espírito Santo"
    | "Goiás"
    | "Maranhão"
    | "Mato Grosso"
    | "Mato Grosso do Sul"
    | "Minas Gerais"
    | "Pará"
    | "Paraíba"
    | "Paraná"
    | "Pernambuco"
    | "Piauí"
    | "Rio de Janeiro"
    | "Rio Grande do Norte"
    | "Rio Grande do Sul"
    | "Rondônia"
    | "Roraima"
    | "Santa Catarina"
    | "São Paulo"
    | "Sergipe"
    | "Tocantins"
    | "TODOS";
  uid:
    | "AC"
    | "AL"
    | "AP"
    | "AM"
    | "BA"
    | "CE"
    | "DF"
    | "ES"
    | "GO"
    | "MA"
    | "MT"
    | "MS"
    | "MG"
    | "PA"
    | "PB"
    | "PR"
    | "PE"
    | "PI"
    | "RJ"
    | "RN"
    | "RS"
    | "RO"
    | "RR"
    | "SC"
    | "SP"
    | "SE"
    | "TO"
    | "all";
}